refactor(architecture): remove init package
