--- website/init/__init__.py
@@ -1,4 +1,0 @@
-from .blueprints import register_blueprints
-from .errors import register_error_handlers
-from .extensions import db, login_manager, mail, limiter, oauth, google, init_markdown
-from .scheduler import schedule_jobs
