+++ website/interface/routes/__init__.py
@@ -1,4 +1,8 @@
 from .auth import auth_bp
 from .admin import admin_bp
+
+from .public import public_bp
 from .settings import settings_bp
+
 from .posts import posts_bp
+from .comments import comments_bp
